Project files/ibm/backend/requirements.txt
fastapi
uvicorn
python-dotenv
pinecone-client
httpx
pydantic
python-multipart
pandas
scikit-learn 